requirements.txt
Django==5.0.3
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
requests==2.31.0
celery==5.3.6
redis==5.0.3
eventlet==0.35.2